[v3.0-rt]OF: Fixup resursive locking code paths
There is no real reason to use a rwlock for devtree_lock. It even
could be a mutex, but unfortunately it's locked from cpu hotplug
pathes which can't schedule :(

So it needs to become a raw lock on rt as well. devtree_lock would be
the only user of a raw_rw_lock, so we are better of cleaning the
recursive locking pathes which allows us to convert devtree_lock to a
read_lock.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>