powerpc/mpic: let IPIs be used as general interrupt sources
On a non-SMP build, the IPIs will be free for other uses, such
as reflecting to a KVM guest.

Signed-off-by: Scott Wood <scottwood@freescale.com>

powerpc/mpic: don't WARN on external IPI usage with SMP

Not all SMP implementations make use of the MPIC IPIs (some
use doorbells, or you could be booting an SMP kernel on a system
with only one CPU visible to this instance of Linux).  So don't
condition the ability to mask the IPI on CONFIG_SMP.

Signed-off-by: Scott Wood <scottwood@freescale.com>