KVM: PPC: e500: eliminate a trap when entering idle
KVM will respond to setting MSR[WE], no matter what is in HID0.

With KVM we also know it will take effect right away, so no need for
a loop, TLF_NAPPING, etc.

KVM paravirt's initcall is moved before SMP init, to avoid any chance
of a race when updating the idle loop handler.  Move this.

Signed-off-by: Scott Wood <scottwood@freescale.com>