KVM: PPC: booke: Add guest debug support
Without hardware support for virtualization, hardware never knows
whither the current pc is in the guest or the host.  So when hardware
single step is enabled for the guest, it cannot be automatically
excluded when entering the host.

Thus, a single step interrupt will happen at the beginning of the
guest exit path.  We need to recognize this kind of single step
interrupt, disable single step, and rfci.

Signed-off-by: Liu Yu <yu.liu@freescale.com>