KVM: PPC: bookehv: Modify linux non base exception handling.
1. Modify the critical exception plolog in order to use ivor_nr
(reuqired for KVM).

2. Modified the EXC_LEVEL_EXCEPTION_PROLOG (shared by
critical/debug/machine check) macro to use DO_KVM.  We intend to
transfer the handling of Debug and guest critical doorbell exceptions
occuring in the guest context to KVM.  When exceptions other than
these occur in the guest context (MSR[GS] = 1), the exception stack
would be used for saving the state irrespective of the MSR[PR] value.

Signed-off-by: Varun Sethi <Varun.Sethi@freescale.com>
Signed-off-by: Scott Wood <scottwood@freescale.com>