kvm: ppc: bookehv: Do away with shadow_pid
Shadow pid is not required in case of e500mc KVM as we use the guest pid
as-is.  We can store the guest pid in arch.pid instead of
arch.shadow_pid.  In this patch we modify references of shadow_pid to pid
for e500mc KVM.

There was a mismatch between pid value stored during GET_REGS and the
value that was restored during SET_REGS.  This resulted in guest pid
corruption, resulting in incorrect tlb entry creation for the guest.

Signed-off-by: Varun Sethi <Varun.Sethi@freescale.com>
Signed-off-by: Scott Wood <scottwood@freescale.com>