drivers/virt: add support for older partman versions to Freescale HV driver
The ioctl interface of original version of the Freescale hypervisor management
driver expected an ioctl type of 0 and a size of at least 40.  This was changed
when the driver was pushed upstream, but we still want to maintain backwards
compatibility with older versions of partman.  If we don't, then everyone
will be forced to rebuild partman if they want to use the latest kernel.

This patch should be applied only to the SDK, and only after the kernel has
been updated to include the upstream versions of the hypervisor device drivers.
This patch should never be posted upstream.

Signed-off-by: Timur Tabi <timur@freescale.com>
Signed-off-by: Kumar Gala <galak@kernel.crashing.org>