KVM: PPC: Support for direct assignment of non-PCI MPIC interrupts
TODO: refactor a common base between this and the PCish PCI interrupt code

I eliminated the workqueue; it is not required for powerpc.  If it is really
needed on x86, then when merging x86 can do the workqueue itself from its
set_irq implementation.

I observed the workqueue indirection adding around 15 us per interrupt,
about 1/3 of the total interrupt latency.

IACK is paravirtualized as as EPR -- this eliminates one MMIO trap per
interrupt.

CTPR is paravirtualized.

Signed-off-by: Scott Wood <scottwood@freescale.com>
Signed-off-by: Stuart Yoder <stuart.yoder@freescale.com>
[stuart: minor merge during rebase]