KVM: PPC: bookehv: Dynamically fixup MSR_GS
Added a new CPU feature CPU_FTR_EMB_HV for e500mc, indicating E.hv support.
This bit get's cleared in the guest context where MMUCFG_LPIDSIZE == 0.
If this bit is cleared the MSR_GS check under DO_KVM is replaced with nops.
This allows the same uImage to be used as both host and guest.

Signed-off-by: Varun Sethi <Varun.Sethi@freescale.com>
Signed-off-by: Laurentiu Tudor <Laurentiu.Tudor@freescale.com>