[v3.0-rt]sched: Do not throttle due to PI boosting
When a runqueue has rt_runtime_us = 0 then the only way it can
accumulate rt_time is via PI boosting. Though that causes the runqueue
to be throttled and replenishing does not change anything due to
rt_runtime_us = 0. So avoid that situation by clearing rt_time and
skip the throttling alltogether.

[ tglx: Preliminary changelog ]

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>