dpaa_eth: Add a few include that are need for building with new kernels
With newer kernel tree (v3.1 or greater) we have build issues because we
can't find the following:

	in_irq
	dma_unmap_single
	dma_map_single
	dma_set_mask
	DMA_BIT_MASK

This is due to the fact that we aren't include the headers that defines
these directly but getting them based on some other include file include
them.

Move to just include <linux/hardirq.h> and <linux/dma-mapping.h> to fix
this issue.

Signed-off-by: Kumar Gala <galak@kernel.crashing.org>