qman: remove support for non-stashing configurations.
Linux no longer needs support for portals whose stashing LIODNs aren't
configured, as there is a native PAMU driver. This commit simplifies
driver interfaces and makes upcoming features easier to implement. In
any case, if a later need for a non-coherent driver mode comes up, it
could be reintroduced in an easier manner than the current code which
is being removed.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>
Signed-off-by: Kumar Gala <galak@kernel.crashing.org>