gianfar: Add BDs allocation in L2 SRAM option
If this option is enabled the driver will try to allocate the
buffer descriptors (BDs) in the L2 SRAM memory, using the
powerpc/85xx cache-sram support API for P1/P2 QorIQ platforms,
with the intent to improve its performance.
For this to work, SRAM size and the offset where SRAM should
be mapped must be provided at kernel command line as:
	cache-sram-size=<value>
	cache-sram-offset=<offset>
Should the attempt to allocate in L2 SRAM fail, the driver falls
back to normal allocation of BDs (i.e. in normal RAM).

Signed-off-by: Sandeep Gopalpet <sandeep.kumar@freescale.com>
Signed-off-by: Jiajun Wu <b06378@freescale.com>
Signed-off-by: Li Yang <leoli@freescale.com>
Signed-off-by: Akhil Goyal <akhil.goyal@freescale.com>
Signed-off-by: Claudiu Manoil <claudiu.manoil@freescale.com>