fmd: fman setup shouldn't use pool-channels
The setup logic for a Fman block was parsing the device-tree looking for
the first ethernet interface bound to that Fman block and parsing out
the corresponding qman pool-channel and using that for scheduling the
result FQ from the Fman. This is wrong for many reasons, but most
important is that the qman driver will soon be fixed to no longer
configure portals to dequeue from all pool-channels by default. As such,
the setup logic for ethernet interfaces will be adjusted to add the
interface's pool-channels to the dequeue masks of the relevant portals,
but this will be after the Fman blocks are set up.

This commit changes the Fman setup logic to just direct Fman setup
results to one of the affine portals directly, thus making no assumption
about pool channels.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>
Acked-by: Stefan Szabo <stefan.szabo@freescale.com>