dpaa_eth: Avoid copying cloned skbs on Tx
In the current driver code, if a cloned skb is received from the stack,
it is unshared (data copy), such that the data buffer can be safely
used to store a backpointer to the skb. This is particularly painful
for large frames.

To avoid copying the entire skb data, leave the skb shared, but
send a S/G frame to Fman, with the S/G table containing a single
entry that points to the skb data buffer. The skb backpointer will
be stored in the first buffer (the one containing the S/G table),
leaving the data buffer untouched.

This improves performance significantly for originating TCP traffic
(~14% for 9.6K jumbo frames).

Signed-off-by: Ioana Radulescu <ruxandra.radulescu@freescale.com>