qbman: export unused portals as UIO devices
This is an overhaul of the previous UIO driver. UIO devices are created
out of any portals that weren't initialised for the kernel's own
requirements, rather than being statically pre-destined for user-space nor
statically pre-configured for stashing affinity to a particular cpu. The
user-space thread should open the device after it has pinned itself to one
cpu, because that's the cpu the driver will set the stashing affinity to.
The code has also been reorganised for better re-use and to get rid of
union structures.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>