fsl_usdpaa: user-space FQID/BPID allocation
This adds ioctl() commands to the fsl_usdpaa character device to allow
(de)allocation of FQIDs and BPIDs. Per-FD accounting is done so cleanup
is automatically performed if deallocation wasn't performed explicitly,
however non-empty pools and non-OOS FQs are intentionally leaked rather
than being deallocated to cause trouble later. (Recover logic will be
able to hook in here later on to correct this.)

Note, the per-FD accounting uses the same dpa_alloc interface to record
the allocated IDs as the qbman driver does to maintain the allocator.
However there is no common "private" header that can be used between
the misc-char driver and the qbman driver, so the definitions are moved to
fsl_usdpaa.h (but protected by __KERNEL__ against user-space use).

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>