dpaa_eth: Add Scatter-Gather support
* Summary *
This patch introduces Scatter-Gather capability in the DPAA-Ethernet
driver. Support is available on both Rx and Tx, forwarding and
termination scenarios.

* Features *
Support is available as a compile-time option, by default turned OFF.
The SG code is mostly all in a new file, conditionally-compiled.
Definitions common to both SG and non-SG code are moved to headers.

This code supports both SG and linear frames/skbuffs, but it is not
optimized for the linear case (if SG support is activated, the linear
frames are processed by code other than the non-SG driver).

On the Rx path, the first 128 bytes of the received frame are memcopied
into the linear part of a new skb. The remaining ingress buffers are
directly used as skb frags.

On the Tx path, the linear part of a nonlinear skb is memcopied to the
first egress SG buffer, while skb frags are used as SG buffers (no copy).
The linear frames are sent without involving a memcopy.

The dynamic FMan resources allocation algorithm can reduce the size
of the Rx FIFOs when Scatter Gather support is enabled. Added the
required modifications to take SG support into account.

This feature is still work in progress, especially in what concerns the
performance improvement. As such, TODOs are still present in the code.

* Known Limitations *
 - No support for NETIF_F_HIGHDMA. This means that non-linear egress
   skbuffs allocated from highmem will be automatically linearized by
   the stack.
 - Only one ingress fragment per page is supported. Buffer pools are
   populated with PAGE_SIZE buffers. An unintended consequence is that
   recycling is not possible with SG code.
 - No integration with ethtool.
 - Performance was not a stated goal in this initial version of the
   feature.

Signed-off-by: Madalin Bucur <madalin.bucur@freescale.com>
Signed-off-by: Ioana Radulescu <ruxandra.radulescu@freescale.com>
Reviewed-by: Bogdan Hamciuc <bogdan.hamciuc@freescale.com>