gianfar: Optimization for routed packets by exchanging rx/tx buffers
Patch enhances the Gianfar driver performance for the packets that are
received by the driver and transmitted by the driver in the same context.
Performance is enhanced by exchanging the skbs between the Tx and Rx rings,
rather than allocating new skbs from kernel or taking it from recycle queue.
The cleanup of BDs is done by the xmit function, so there is no need of the
Tx interrupts, or Tx cleanup functionality in the receive path, avoiding the
locks for cleanups

Patch gives best performance with DEFAULT_RX_RING_SIZE 32, DEFAULT_TX_RING_SIZE 32,
and GFAR_DEV_WEIGHT 32

Signed-off-by: Rajan Gupta <rajan.gupta@freescale.com>