gianfar: Updating transmit time in tx queue in gianfar driver
In RX_TX_BUFF_XCHG flag transmit queue is selected in the
gfar_start_xmit function based on the processor id to avoid
taking lock while transmitting the packet.
As transmit queue selected in driver, networking stack is
not updating the time stamp for the tramsmit queue.
Updating the time stamp in transmit queue in gfar_start_xmit
and gfar_fast_xmit functions.

Signed-off-by: Rajan Gupta <rajan.gupta@freescale.com>