qman: Fix race condition in congestion state change handler
QMan's ISR handler must clear the CSCI bit *before* interrogating the
hardware for the current congestion status. Otherwise, a
quickly-switching state might leave the CGR indefinitely congested.

An example to this scenario is when the CGSCN handler is relatively
slow, while the CGR's congestion threshold is low enough to set the
CSCI bit in a very quick sequence.

Clearing the CSCI bit inside __poll_portal_slow() to avoid the said
race.

Signed-off-by: Bogdan Hamciuc <bogdan.hamciuc@freescale.com>
Signed-off-by: Alex Marginean <alexandru.marginean@freescale.com>