fmd: remove support for loading the microcode from NOR flash
Early versions of FMD read the Fman microcode from NOR flash, but that feature
has been deprecated for quite some time.  The feature is not normally
compiled anyway, but we don't want anyone to think that it's an acceptable
approach.  The only supported method for obtaining the microcode is from
the device tree.

Signed-off-by: Timur Tabi <timur@freescale.com>
Signed-off-by: Stefan Szabo <szbs001@freescale.com>