net: Make the netdev watchdog aware of hardware multiqueue devices
If the netdev declares the NETIF_F_HW_ACCEL_MQ (accelerated multiqueue)
capability, tell the watchdog to consider the per-netdev trans_start
field rather than any individual multiqueue's timestamp. That is
justified by the fact that queues only go in and out of congestion
in groups, not individually, as far as the net device is concerned.

Signed-off-by: Bogdan Hamciuc <bogdan.hamciuc@freescale.com>