powerpc/uio: Add uio driver for rapidio message and doorbell units
For using the rapidio message and doorbell units in linux user space,
need to build the uio devices in kernel space. The message and
doorbell uio devices in kernel will map the units registers to user
space and then the applications in user space can configure those
registers and use the rapidio message and doorbell units.

Signed-off-by: Liu Gang <Gang.Liu@freescale.com>