dpaa_eth: Selectively disable GRO
For certain types of traffic and/or PCD configurations, GRO is
undesirable because of unnecessary extra header processing. This becomes
much more of a problem with small frames (e.g. 64-byte UDP, or 64-byte
IP) and going through GRO when we know for sure it isn't the case incurs
a heavy performance penalty.

Signed-off-by: Bogdan Hamciuc <bogdan.hamciuc@freescale.com>