dpaa_eth: Fix initial value of netdev MTU
If not overridden at early initialization of the net device, the netdev
receives a default MTU of 1500 which can be changed later via ifconfig,
respectively dev_set_mtu().

In cases where we boot the kernel with fsl_fm_max_frm below 1522, we
shouldn't see, or be allowed to set, MTU 1500 on any of the dpaa-eth net
devices. But dev_set_mtu() will not invoke the driver's .ndo_change_mtu
callback if the "new" MTU is the same as the "old" MTU. This results in
a logical crack, through which we can slip an MTU=1500 which is outside
our self-defined range.

Signed-off-by: Bogdan Hamciuc <bogdan.hamciuc@freescale.com>