fmd: Fix fmlib FM_VSP_Config wrapper function for FM object member
When FM_VSP_Config is called from userspace, h_Fm parameter member
transmited to the function is not mapped in kernel with the Fm object.
h_Fm is a descriptor returned in userspace by FM_Open fmlib wrapper function.
Thus when copying this param from user to kernel the value transmited is actually
a descriptor instead of the FMAN handle. This patch fixes the above issue,
transmitting the FMAN handle to the FM_VSP_Config from FMD driver.

Signed-off-by: Radu Bulie <radu.bulie@freescale.com>
Acked-by: Madalin Bucur <madalin.bucur@freescale.com>