fmd: Minor change to FMD wrapper API
The FMD linux wrapper offers some API functions for the use of the
DPAA Ethernet / DPAA Offline port driver, which must respect the
Linux naming style conventions.

Modify a parameter related to the initialization of OH ports for
IP fragmentation which was not following the Linux style.
This is a non-functional change.

Signed-off-by: Ioana Radulescu <ruxandra.radulescu@freescale.com>