dpaa_eth: Make header reallocation unlikely
The headroom of an egress skb should be at least equal to
the netdevice's needed_headroom, even in forwarding cases. In fact,
the stack itself reallocs the headroom (ip_finish_output2()) if the
current one is less than LL_RESERVED_SPACE(dev).

Make the guarding condition unlikely. Probably the only scenario we're
talking about is ASF, and even that only in some cases.

Signed-off-by: Bogdan Hamciuc <bogdan.hamciuc@freescale.com>