qman: Increase default size of PFDR pool
In (not so) pathological scenarios where cores are transmitting small
datagrams at a very fast pace, QMan's PFDR pool may deplete and a "PFDR
low" interrupt is raised, shortly followed by an "enqueues blocked
interrupt".
Because the combination of the two may lead to a system-wide deadlock
(all entities - cores, FMan - remaining stuck while trying to
enqueue to a PFDR-depleted QMan), the worst-case scenario may entail a
system reboot.

As it turns out, increasing the default PFDR pool size from 16M to 32M
suffices to work around the described problem, at least on a 8-core
P4080 (on which the problem was reported).

While we're at it, moved the hard-coded DEFAULT_PFDR_SZ value to the
Kconfig.

Signed-off-by: Bogdan Hamciuc <bogdan.hamciuc@freescale.com>
Signed-off-by: Andy Fleming <afleming@freescale.com>