dpaa_eth: Adjust congestion threshold for 1G ports
Small-sized egress datagrams present a memory management problem,
because:
  1) They can be sent in very tight loops (no protocol-enforced rate
  limiting);
  2) They carry a fixed - but significant - overhead from skb shells and
  the netdevice's needed_headroom, which is difficult to take into
  account with the current DPAA support.

As such, high enqueueing rates of UDP datagrams sized 512 and below may
eventually lead to OOM, unless the Tx rate is throttled. CGRs are the
mechanism-of-choice to do that, but the value previously set for the
congestion threshold was too high and in some cases it wasn't hit before
memory was depleted.

Lowering the congestion threshold for 1Gbps ports to a value
experimentally found to be attained, such that transmission is stopped
before the system runs out of memory.

For datagram-based traffic sized 512 and above, this incurs a ~2%
bandwidth loss on 1Gbps ports. No penalty is expected on 10Gbps ports.

Signed-off-by: Bogdan Hamciuc <bogdan.hamciuc@freescale.com>