dpaa_eth: Optionally disable ndo_select_queue in favor of XPS
The existence of a ndo_select_queue() callback overrides the XPS
selection logic on the egress path. Provide a Kconfig choice of
disabling this callback, to allow the standard XPS selection logic and
user-controlled FQ-to-CPU mapping.

In addition, if XPS is chosen, use the queue from
skb->queue_mapping for transmission.

Signed-off-by: Bogdan Hamciuc <bogdan.hamciuc@freescale.com>