powerpc/mpc85xx/e6500: module support for A-006198 workaround
By letting the compiler obtain the address of the function, we should
avoid problems with module relocation.  Note that on the ppc64 ABI,
when we convert a function pointer to an integer pointer, we get a
pointer to a descriptor rather than the program code, and thus we
need an extra dereference compared to other ABIs.  It is likely that
the relocation problem resulted from the previous asm code trying to
directly reference the code address (i.e. the symbol beginning with a
period).

Signed-off-by: Scott Wood <scottwood@freescale.com>
Signed-off-by: Andy Fleming <afleming@freescale.com>