dpaa_eth: Remove congestion group initialization for mac-less interfaces
MAC-less devices initialize their own RX frame queues, assuming that
the TX queues will be initialized by the other partition.

We cannot rely on the second partition to place the TX queues of
the first partition into congestion groups because the congestion
interrupt will then be received by the second partition instead
of the partition enqueueing the frames (in this case the first
partition).

Supporting congestion on MAC-less devices would involve too much
code and device tree changes so we simply use taildrop mechanism
for now.

Signed-off-by: Cristian Sovaiala <cristian.sovaiala@freescale.com>