powerpc/85xx: Enabled stashing of FMAN write data.
This patch enables translation of FMAN write transactions into
write-with-stashing (to CPC) using IOMMU operation mapping mechanism.
This is required for USDPAA IPSEC application so that SEC is able to DMA
data faster and show better performance. On T4240, this patch shows up
20% performance gain for USDPAA IPSEC.

The stashing of FMAN write data is controllable with a kernel
configuration option.

Signed-off-by: Vakul Garg <vakul@freescale.com>