dpa_offload: Reduce the size of the IPv4 header options field
According to RFC 791, the size of the IPv4 header (options
included) is represented on 4 bits (IHL), hence it can be an
absolute maximum of 15 * 4 = 60 bytes. The IPv4 options buffer
size field in the fsl_dpa_offload.h header file, in the
ipv4_header data structure, was declared as an "unsigned int" and
this was causing a static analysis warning in dpa_classifier.

The field was reduced to uint8_t which is more than enough to
store the size of the IPv4 frame options.

Signed-off-by: Marian Chereji <marian.chereji@freescale.com>
Change-Id: I69fb1307329b68da652ed85d659aa10e34c5af67
Reviewed-on: http://git.am.freescale.net:8181/4264
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Bulie Radu-Andrei-B37577 <Radu.Bulie@freescale.com>
Reviewed-by: Rivera Jose-B46482 <German.Rivera@freescale.com>