dpa_offload: Fix IPv4 options size field inconsistency
Due to a bad sync the ipv4_header defines the attribute options_size as
uint8_t, while the compat_ipv4_header defines it as unsigned int. The
compat_ipv4_header must define this attribute as uint8_t as well to
avoid problems when casting from one structure to the other.

Signed-off-by: Marian Chereji <marian.chereji@freescale.com>
Change-Id: I98df663fda9d0ea19f0e17ac71f99512b06c5bb6
Reviewed-on: http://git.am.freescale.net:8181/5742
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Richard Schmitt <B43082@freescale.com>