dpa_offload: Improve completion message when Modify SA function is called
Before this patch the modify SA descriptor is setting only the SA id
in the output frame, this was enough when  only the Anti Replay Window
could be modified. If adding the other types of change line SEQ_NUM or
EXT_SEQ_NUM than the SA id is insufficient because the user
will not know what operation has finished on the SA with the ID specified in the frame.

Improvement: an operation code and the SA id are added inside the output FD
created by SEC engine. The code is highligting the operation
that took place on the SA with that ID.

Signed-off-by: Andrei Varvara <andrei.varvara@freescale.com>
Change-Id: I62e5497422797257a4d5e658703658b278f553b4
Reviewed-on: http://git.am.freescale.net:8181/6036
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Marian-Cornel Chereji <marian.chereji@freescale.com>
Reviewed-by: Thomas Trefny <Tom.Trefny@freescale.com>