dpa_offload: Add support for getting the current SA Sequence Number or ESN
The requirements is needed in order to support IPSec High Availability.

Add new functions into the dpa_ipsec API:
    -dpa_ipsec_sa_request_seq_number
    -dpa_ipsec_sa_get_seq_number

Request the sequence number of an SA asynchronous - SEC will
dequeue a frame with RJD, run it and after this create an
output frame with status of user error. The frame will have always the
length of 5 bytes, first one representing the operation code that has
finished and the next 4 will determine the SA id on which the operation took
place.

The get seq number have to be called by the user when it receives
the fd with user error for this SA. The API will return the SEQ
number.

Signed-off-by: Andrei Varvara <andrei.varvara@freescale.com>
Change-Id: I154ea8ee9cfc95eae0ef24cdb3f3559a0748baf0
Reviewed-on: http://git.am.freescale.net:8181/6037
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Marian-Cornel Chereji <marian.chereji@freescale.com>
Reviewed-by: Thomas Trefny <Tom.Trefny@freescale.com>